+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- New hook: `Turboprop.Hooks.Dialog`
+
 ## 0.1.4 - 2024.06-11
 
 ### Added